[RL support matrix] Refactors the DP group sampling with prefix caching tests into a standalone test (#2641)
