+++ .buildkite/rl/group_sampling_prefix_cache.yml
@@ -1,0 +1,68 @@
+# Copyright 2025 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+# Tests for group sampling (n=16) with prefix caching: verifies that repeated
+# requests with the same prompt produce diverse outputs and benefit from the
+# prefix cache.
+steps:
+  # ------------------------------------------------------------------
+  # Correctness tests
+  # ------------------------------------------------------------------
+  - label: "${TPU_VERSION:-tpu6e} Correctness tests for group_sampling_prefix_cache"
+    key: "${TPU_VERSION:-tpu6e}_group_sampling_prefix_cache_CorrectnessTest"
+    soft_fail: true
+    if: build.env("NIGHTLY") == "1"
+    agents:
+      queue: "${TPU_QUEUE_MULTI:-tpu_v6e_8_queue}"
+    commands:
+      - .buildkite/scripts/run_in_docker.sh python3 -m pytest -s -v /workspace/tpu_inference/tests/core/test_dp_group_sampling_prefix_cache.py::TestGroupSamplingPrefixCache::test_group_sampling_produces_diverse_outputs
+  - label: "${TPU_VERSION:-tpu6e} Record correctness test result for group_sampling_prefix_cache"
+    key: "${TPU_VERSION:-tpu6e}_record_group_sampling_prefix_cache_CorrectnessTest"
+    depends_on: "${TPU_VERSION:-tpu6e}_group_sampling_prefix_cache_CorrectnessTest"
+    env:
+      CI_TPU_VERSION: "${TPU_VERSION:-tpu6e}"
+      CI_TARGET: "Group sampling prefix cache"
+      CI_STAGE: "CorrectnessTest"
+      CI_CATEGORY: "rl support matrix"
+    agents:
+      queue: cpu
+    commands:
+      - |
+        .buildkite/scripts/record_step_result.sh ${TPU_VERSION:-tpu6e}_group_sampling_prefix_cache_CorrectnessTest
+
+  # ------------------------------------------------------------------
+  # Performance tests
+  # ------------------------------------------------------------------
+  - label: "${TPU_VERSION:-tpu6e} Performance tests for group_sampling_prefix_cache"
+    key: "${TPU_VERSION:-tpu6e}_group_sampling_prefix_cache_PerformanceTest"
+    depends_on: "${TPU_VERSION:-tpu6e}_record_group_sampling_prefix_cache_CorrectnessTest"
+    soft_fail: true
+    if: build.env("NIGHTLY") == "1"
+    agents:
+      queue: "${TPU_QUEUE_MULTI:-tpu_v6e_8_queue}"
+    commands:
+      - .buildkite/scripts/run_in_docker.sh python3 -m pytest -s -v /workspace/tpu_inference/tests/core/test_dp_group_sampling_prefix_cache.py::TestGroupSamplingPrefixCache::test_group_sampling_hits_prefix_cache
+  - label: "${TPU_VERSION:-tpu6e} Record performance test result for group_sampling_prefix_cache"
+    key: "${TPU_VERSION:-tpu6e}_record_group_sampling_prefix_cache_PerformanceTest"
+    depends_on: "${TPU_VERSION:-tpu6e}_group_sampling_prefix_cache_PerformanceTest"
+    env:
+      CI_TPU_VERSION: "${TPU_VERSION:-tpu6e}"
+      CI_TARGET: "Group sampling prefix cache"
+      CI_STAGE: "PerformanceTest"
+      CI_CATEGORY: "rl support matrix"
+    agents:
+      queue: cpu
+    commands:
+      - |
+        .buildkite/scripts/record_step_result.sh ${TPU_VERSION:-tpu6e}_group_sampling_prefix_cache_PerformanceTest
